refactor(bazed-svelte): rename Rekuho to Bazed
